+++ lib/lambda_runtime.mli
@@ -119,6 +119,7 @@ module Json : sig
        and type response := Yojson.Safe.t
 end
 
+(** API Gateway v1 event and response definitions *)
 module Http : sig
   open Util
 
@@ -156,6 +157,7 @@ module Http : sig
     ; api_id : string  (** The API Gateway REST API ID *)
     }
 
+  (* A request to API Gateway using the proxy integration *)
   type api_gateway_proxy_request =
     { resource : string
     ; path : string
@@ -167,6 +169,75 @@ module Http : sig
     ; request_context : api_gateway_proxy_request_context
     ; body : string option
     ; is_base64_encoded : bool
+    }
+
+  type api_gateway_proxy_response =
+    { status_code : int
+    ; headers : string StringMap.t
+    ; body : string
+    ; is_base64_encoded : bool
+    }
+
+  include
+    LambdaRuntime
+      with type event := api_gateway_proxy_request
+       and type response := api_gateway_proxy_response
+end
+
+(** API Gateway v2 event and response definitions *)
+module Httpv2 : sig
+  open Util
+
+  type api_gateway_proxy_request_context_http =
+    { method_ : string
+    ; path : string
+    ; protocol : string
+    ; source_ip : string
+    ; user_agent : string
+    }
+
+  type api_gateway_request_context_jwt =
+    { claims : string StringMap.t [@default StringMap.empty]
+    ; scopes : string StringMap.t [@default StringMap.empty]
+    }
+
+  type api_gateway_proxy_request_context_authorizer =
+    { jwt : api_gateway_request_context_jwt }
+
+  (* APIGatewayProxyRequestContext contains the information to identify the AWS
+   * account and resources invoking the Lambda function. It also includes Cognito
+   * identity information for the caller. *)
+  type api_gateway_proxy_request_context =
+    { account_id : string
+    ; api_id : string
+    ; domain_name : string
+    ; domain_prefix : string
+    ; http : api_gateway_proxy_request_context_http
+    ; resource_id : string option
+    ; stage : string
+    ; request_id : string
+    ; route_key : string
+    ; time : string
+    ; time_epoch : int64
+    ; authorizer : api_gateway_proxy_request_context_authorizer option
+    }
+
+  (* A request to API Gateway using the proxy integration (v2).
+
+     This differs from Http.api_gateway_proxy_request in that it does away with
+     multi_value_headers *)
+  type api_gateway_proxy_request =
+    { version : string
+    ; route_key : string
+    ; raw_query_string : string
+    ; cookies : string list option
+    ; headers : string StringMap.t
+    ; query_string_parameters : string StringMap.t
+    ; request_context : api_gateway_proxy_request_context
+    ; body : string option
+    ; path_parameters : string StringMap.t
+    ; is_base64_encoded : bool
+    ; stage_variables : string StringMap.t
     }
 
   type api_gateway_proxy_response =