+++ CHANGES.md
@@ -4,6 +4,7 @@ Unreleased
 - Tolerate unknown fields when parsing JSON payloads
   (`@@deriving yojson {strict = false; }`)
   ([#62](https://github.com/anmonteiro/aws-lambda-ocaml-runtime/pull/62))
+- Add API Gateway v2 definitions ([#68](https://github.com/anmonteiro/aws-lambda-ocaml-runtime/pull/68))
 
 0.1.0 2021-04-17
 --------------